feat(ux): license banners across apps + action gating on EXPIRED
