+++ packages/api/src/endpoints.test.ts
@@ -1,5 +1,4 @@
-import { strict as assert } from 'node:assert';
-import { test, afterEach } from 'node:test';
+import { describe, test, afterEach, expect } from 'vitest';
 import { exportMenuI18n } from './endpoints';
 
 afterEach(() => {
@@ -7,14 +6,16 @@ afterEach(() => {
   delete globalThis.fetch;
 });
 
-test('exportMenuI18n returns CSV with expected columns', async () => {
-  const csv = 'id,en_name,hi_name\n';
-  // @ts-ignore
-  globalThis.fetch = async (url: RequestInfo | URL, init?: RequestInit) => {
-    assert.equal(url, '/menu/i18n/export?lang=en&lang=hi');
-    assert.deepEqual(init, { headers: {} });
-    return new Response(csv, { status: 200 });
-  };
-  const res = await exportMenuI18n(['en', 'hi']);
-  assert.equal(res, csv);
+describe('exportMenuI18n', () => {
+  test('returns CSV with expected columns', async () => {
+    const csv = 'id,en_name,hi_name\n';
+    // @ts-ignore
+    globalThis.fetch = async (url: RequestInfo | URL, init?: RequestInit) => {
+      expect(url).toBe('/menu/i18n/export?lang=en&lang=hi');
+      expect(init).toEqual({ headers: {} });
+      return new Response(csv, { status: 200 });
+    };
+    const res = await exportMenuI18n(['en', 'hi']);
+    expect(res).toBe(csv);
+  });
 });
