+++ packages/ui/src/components/index.ts
@@ -2,4 +2,4 @@ export * from './button';
 export * from './global-error-boundary';
 export * from './Skeleton';
 export * from './empty-state';
-export * from './license-banner';
+export * from './LicenseBanner';
